Streamline system nexus signal-with-start setup
