+++ internal/workflow.go
@@ -3,7 +3,6 @@ package internal
 import (
 	"cmp"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"reflect"
@@ -11,10 +10,9 @@ import (
 	"strings"
 	"time"
 
+	"github.com/google/uuid"
 	"github.com/nexus-rpc/sdk-go/nexus"
 
-	"go.temporal.io/api/temporalproto"
-	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/durationpb"
 
 	commonpb "go.temporal.io/api/common/v1"
@@ -1791,9 +1789,7 @@ func SignalWithStartWorkflow(
 
 	options.ID = workflowID
 	if options.ID == "" {
-		future, settable := NewFuture(ctx)
-		settable.Set(nil, errWorkflowIDNotSet)
-		return future
+		options.ID = uuid.NewString()
 	}
 
 	workflowType, err := getWorkflowFunctionName(env.GetRegistry(), workflow)
@@ -1819,8 +1815,9 @@ func (wc *workflowEnvironmentInterceptor) SignalWithStartWorkflow(
 	ctx1 := setWorkflowEnvOptionsIfNotExist(ctx)
 	future, settable := NewFuture(ctx1)
 
-	params, err := wc.prepareSignalWithStartNexusOperationParams(
+	params, err := prepareSystemNexusSignalWithStartOperationParams(
 		ctx1,
+		wc.env,
 		workflowID,
 		signalName,
 		signalArg,
@@ -1899,397 +1896,6 @@ func signalExternalWorkflow(ctx Context, workflowID, runID, signalName string, a
 	)
 
 	return future
-}
-
-func (wc *workflowEnvironmentInterceptor) prepareSignalWithStartNexusOperationParams(
-	ctx Context,
-	workflowID string,
-	signalName string,
-	signalArg interface{},
-	options StartWorkflowOptions,
-	workflowType string,
-	workflowArgs []interface{},
-) (executeNexusOperationParams, error) {
-	if workflowID == "" {
-		return executeNexusOperationParams{}, errWorkflowIDNotSet
-	}
-
-	workflowOptionsFromCtx := getWorkflowEnvOptions(ctx)
-	dc := WithWorkflowContext(ctx, workflowOptionsFromCtx.DataConverter)
-	env := getWorkflowEnvironment(ctx)
-	wfType, input, err := getValidatedWorkflowFunction(workflowType, workflowArgs, dc, env.GetRegistry())
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-
-	signalInput, err := encodeArg(dc, signalArg)
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-
-	header, err := workflowHeaderPropagated(ctx, workflowOptionsFromCtx.ContextPropagators)
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-
-	memo, err := getWorkflowMemo(options.Memo, dc, env.TryUse(SDKFlagMemoUserDCEncode))
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-
-	searchAttr, err := serializeSearchAttributes(options.SearchAttributes, options.TypedSearchAttributes)
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-
-	userMetadata, err := buildUserMetadata(options.StaticSummary, options.StaticDetails, dc)
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-	var userMetadataMessage proto.Message
-	if userMetadata != nil {
-		userMetadataMessage = userMetadata
-	}
-
-	req, err := newSystemNexusSignalWithStartInput(
-		workflowOptionsFromCtx.Namespace,
-		"",
-		workflowID,
-		signalName,
-		wfType,
-		input,
-		signalInput,
-		header,
-		memo,
-		searchAttr,
-		userMetadataMessage,
-		options,
-	)
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-
-	outerPayload, err := getSystemNexusPayloadConverter().ToPayload(req)
-	if err != nil {
-		return executeNexusOperationParams{}, err
-	}
-
-	return executeNexusOperationParams{
-		client: nexusClient{
-			endpoint: systemNexusEndpoint,
-			service:  systemnexus.WorkflowService.ServiceName,
-		},
-		operation:   systemnexus.WorkflowService.SignalWithStartWorkflowExecution.Name(),
-		input:       outerPayload,
-		options:     NexusOperationOptions{CancellationType: NexusOperationCancellationTypeWaitCompleted},
-		nexusHeader: nexus.Header{},
-	}, nil
-}
-
-func newSystemNexusSignalWithStartInput(
-	namespace string,
-	requestID string,
-	workflowID string,
-	signalName string,
-	workflowType *WorkflowType,
-	input *commonpb.Payloads,
-	signalInput *commonpb.Payloads,
-	header *commonpb.Header,
-	memo *commonpb.Memo,
-	searchAttr *commonpb.SearchAttributes,
-	userMetadata proto.Message,
-	options StartWorkflowOptions,
-) (systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput, error) {
-	req := systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{
-		Namespace:  namespace,
-		RequestID:  requestID,
-		WorkflowID: workflowID,
-		SignalName: signalName,
-		TaskQueue:  systemNexusTaskQueue(options.TaskQueue),
-		WorkflowType: &systemnexus.WorkflowType{
-			Name: workflowType.Name,
-		},
-		WorkflowIDConflictPolicy: systemNexusWorkflowIDConflictPolicy(options.WorkflowIDConflictPolicy),
-		WorkflowIDReusePolicy:    systemNexusWorkflowIDReusePolicy(options.WorkflowIDReusePolicy),
-		VersioningOverride:       systemNexusVersioningOverride(options.VersioningOverride),
-		Priority:                 systemNexusPriority(options.Priority),
-	}
-	var err error
-	if req.RetryPolicy, err = systemNexusRetryPolicy(options.RetryPolicy); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	}
-
-	if req.Input, err = systemNexusInputFromPayloads(input); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	}
-	if req.SignalInput, err = systemNexusInputFromPayloads(signalInput); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	}
-	if req.Header, err = systemNexusHeaderFromProto(header); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	}
-	if req.Memo, err = systemNexusMemoFromProto(memo); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	}
-	if req.SearchAttributes, err = systemNexusSearchAttributesFromProto(searchAttr); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	}
-	if req.UserMetadata, err = systemNexusUserMetadataFromProto(userMetadata); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	}
-
-	if executionTimeout, err := systemNexusDurationString(options.WorkflowExecutionTimeout); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	} else {
-		req.WorkflowExecutionTimeout = executionTimeout
-	}
-	if runTimeout, err := systemNexusDurationString(options.WorkflowRunTimeout); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	} else {
-		req.WorkflowRunTimeout = runTimeout
-	}
-	if taskTimeout, err := systemNexusDurationString(options.WorkflowTaskTimeout); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	} else {
-		req.WorkflowTaskTimeout = taskTimeout
-	}
-	if startDelay, err := systemNexusDurationString(options.StartDelay); err != nil {
-		return systemnexus.WorkflowServiceSignalWithStartWorkflowExecutionInput{}, err
-	} else {
-		req.WorkflowStartDelay = startDelay
-	}
-
-	req.CronSchedule = options.CronSchedule
-	return req, nil
-}
-
-func systemNexusInputFromPayloads(payloads *commonpb.Payloads) (*systemnexus.Input, error) {
-	if payloads == nil || len(payloads.Payloads) == 0 {
-		return nil, nil
-	}
-	value, err := systemNexusJSONValue(payloads)
-	if err != nil {
-		return nil, err
-	}
-	valueMap, ok := value.(map[string]any)
-	if !ok {
-		return nil, errors.New("system nexus payloads JSON must be an object")
-	}
-	items, _ := valueMap["payloads"].([]any)
-	return &systemnexus.Input{Payloads: items}, nil
-}
-
-func systemNexusHeaderFromProto(header *commonpb.Header) (*systemnexus.Header, error) {
-	if header == nil {
-		return nil, nil
-	}
-	value, err := systemNexusJSONValue(header)
-	if err != nil {
-		return nil, err
-	}
-	valueMap, ok := value.(map[string]any)
-	if !ok {
-		return nil, errors.New("system nexus header JSON must be an object")
-	}
-	fields, _ := valueMap["fields"].(map[string]any)
-	return &systemnexus.Header{Fields: fields}, nil
-}
-
-func systemNexusMemoFromProto(memo *commonpb.Memo) (*systemnexus.Memo, error) {
-	if memo == nil {
-		return nil, nil
-	}
-	value, err := systemNexusJSONValue(memo)
-	if err != nil {
-		return nil, err
-	}
-	valueMap, ok := value.(map[string]any)
-	if !ok {
-		return nil, errors.New("system nexus memo JSON must be an object")
-	}
-	fields, _ := valueMap["fields"].(map[string]any)
-	return &systemnexus.Memo{Fields: fields}, nil
-}
-
-func systemNexusSearchAttributesFromProto(searchAttr *commonpb.SearchAttributes) (*systemnexus.SearchAttributes, error) {
-	if searchAttr == nil {
-		return nil, nil
-	}
-	value, err := systemNexusJSONValue(searchAttr)
-	if err != nil {
-		return nil, err
-	}
-	valueMap, ok := value.(map[string]any)
-	if !ok {
-		return nil, errors.New("system nexus search attributes JSON must be an object")
-	}
-	fields, _ := valueMap["indexedFields"].(map[string]any)
-	return &systemnexus.SearchAttributes{IndexedFields: fields}, nil
-}
-
-func systemNexusUserMetadataFromProto(userMetadata proto.Message) (*systemnexus.UserMetadata, error) {
-	if userMetadata == nil {
-		return nil, nil
-	}
-	value, err := systemNexusJSONValue(userMetadata)
-	if err != nil {
-		return nil, err
-	}
-	valueMap, ok := value.(map[string]any)
-	if !ok {
-		return nil, errors.New("system nexus user metadata JSON must be an object")
-	}
-	return &systemnexus.UserMetadata{
-		Summary: valueMap["summary"],
-		Details: valueMap["details"],
-	}, nil
-}
-
-func systemNexusJSONValue(message proto.Message) (any, error) {
-	data, err := temporalproto.CustomJSONMarshalOptions{
-		Metadata: map[string]interface{}{
-			commonpb.EnablePayloadShorthandMetadataKey: true,
-		},
-	}.Marshal(message)
-	if err != nil {
-		return nil, err
-	}
-
-	var value any
-	if err := json.Unmarshal(data, &value); err != nil {
-		return nil, err
-	}
-	return value, nil
-}
-
-func systemNexusDurationString(d time.Duration) (string, error) {
-	if d == 0 {
-		return "", nil
-	}
-	value, err := systemNexusJSONValue(durationpb.New(d))
-	if err != nil {
-		return "", err
-	}
-	durationValue, ok := value.(string)
-	if !ok {
-		return "", errors.New("system nexus duration JSON must be a string")
-	}
-	return durationValue, nil
-}
-
-func systemNexusTaskQueue(name string) *systemnexus.TaskQueue {
-	if name == "" {
-		return nil
-	}
-	kind := systemnexus.TaskQueueKindNormal
-	return &systemnexus.TaskQueue{Name: name, Kind: &kind}
-}
-
-func systemNexusRetryPolicy(retryPolicy *RetryPolicy) (*systemnexus.RetryPolicy, error) {
-	if retryPolicy == nil {
-		return nil, nil
-	}
-	policy := &systemnexus.RetryPolicy{
-		BackoffCoefficient:     retryPolicy.BackoffCoefficient,
-		MaximumAttempts:        int64(retryPolicy.MaximumAttempts),
-		NonRetryableErrorTypes: retryPolicy.NonRetryableErrorTypes,
-	}
-	if retryPolicy.InitialInterval != 0 {
-		initialInterval, err := systemNexusDurationString(retryPolicy.InitialInterval)
-		if err != nil {
-			return nil, err
-		}
-		policy.InitialInterval = initialInterval
-	}
-	if retryPolicy.MaximumInterval != 0 {
-		maximumInterval, err := systemNexusDurationString(retryPolicy.MaximumInterval)
-		if err != nil {
-			return nil, err
-		}
-		policy.MaximumInterval = maximumInterval
-	}
-	return policy, nil
-}
-
-func systemNexusWorkflowIDConflictPolicy(policy enumspb.WorkflowIdConflictPolicy) *systemnexus.WorkflowIDConflictPolicy {
-	switch policy {
-	case enumspb.WORKFLOW_ID_CONFLICT_POLICY_FAIL:
-		value := systemnexus.WorkflowIDConflictPolicyFail
-		return &value
-	case enumspb.WORKFLOW_ID_CONFLICT_POLICY_USE_EXISTING:
-		value := systemnexus.WorkflowIDConflictPolicyUseExisting
-		return &value
-	case enumspb.WORKFLOW_ID_CONFLICT_POLICY_TERMINATE_EXISTING:
-		value := systemnexus.WorkflowIDConflictPolicyTerminateExisting
-		return &value
-	default:
-		return nil
-	}
-}
-
-func systemNexusWorkflowIDReusePolicy(policy enumspb.WorkflowIdReusePolicy) *systemnexus.WorkflowIDReusePolicy {
-	switch policy {
-	case enumspb.WORKFLOW_ID_REUSE_POLICY_ALLOW_DUPLICATE:
-		value := systemnexus.WorkflowIDReusePolicyAllowDuplicate
-		return &value
-	case enumspb.WORKFLOW_ID_REUSE_POLICY_ALLOW_DUPLICATE_FAILED_ONLY:
-		value := systemnexus.WorkflowIDReusePolicyAllowDuplicateFailedOnly
-		return &value
-	case enumspb.WORKFLOW_ID_REUSE_POLICY_REJECT_DUPLICATE:
-		value := systemnexus.WorkflowIDReusePolicyRejectDuplicate
-		return &value
-	case enumspb.WORKFLOW_ID_REUSE_POLICY_TERMINATE_IF_RUNNING:
-		value := systemnexus.WorkflowIDReusePolicyTerminateIfRunning
-		return &value
-	default:
-		return nil
-	}
-}
-
-func systemNexusVersioningOverride(versioningOverride VersioningOverride) *systemnexus.VersioningOverride {
-	if versioningOverride == nil {
-		return nil
-	}
-	switch v := versioningOverride.(type) {
-	case *PinnedVersioningOverride:
-		behavior := systemnexus.VersioningBehaviorPinned
-		pinnedBehavior := systemnexus.PinnedOverrideBehaviorPinned
-		return &systemnexus.VersioningOverride{
-			Behavior:      &behavior,
-			PinnedVersion: v.Version.toCanonicalString(),
-			Deployment: &systemnexus.Deployment{
-				SeriesName: v.Version.DeploymentName,
-				BuildID:    v.Version.BuildID,
-			},
-			Pinned: &systemnexus.Pinned{
-				Behavior: &pinnedBehavior,
-				Version: &systemnexus.Version{
-					DeploymentName: v.Version.DeploymentName,
-					BuildID:        v.Version.BuildID,
-				},
-			},
-		}
-	case *AutoUpgradeVersioningOverride:
-		behavior := systemnexus.VersioningBehaviorAutoUpgrade
-		return &systemnexus.VersioningOverride{
-			Behavior:    &behavior,
-			AutoUpgrade: true,
-		}
-	default:
-		return nil
-	}
-}
-
-func systemNexusPriority(priority Priority) *systemnexus.Priority {
-	var defaultPriority Priority
-	if priority == defaultPriority {
-		return nil
-	}
-	return &systemnexus.Priority{
-		PriorityKey:    int64(priority.PriorityKey),
-		FairnessKey:    priority.FairnessKey,
-		FairnessWeight: float64(priority.FairnessWeight),
-	}
 }
 
 // UpsertSearchAttributes is used to add or update workflow search attributes.
@@ -3469,7 +3075,7 @@ func (wc *workflowEnvironmentInterceptor) prepareNexusOperationParams(ctx Contex
 
 	payloadConverter := dc
 	if systemnexus.IsTemporalNexusOperation(input.Client.Service(), operationName) {
-		payloadConverter = getSystemNexusPayloadConverter()
+		payloadConverter = converter.GetDefaultDataConverter()
 	}
 
 	payload, err := payloadConverter.ToPayload(input.Input)
